+++ tests/test_tools_pdf2txt.py
@@ -2,6 +2,7 @@ import filecmp
 import os
 from shutil import rmtree
 from tempfile import mkdtemp
+from typing import List
 
 from tests.helpers import absolute_sample_path
 from tests.tempfilepath import TemporaryFilePath
@@ -132,7 +133,7 @@ class TestPdf2Txt:
 
 class TestDumpImages:
     @staticmethod
-    def extract_images(input_file, *args):
+    def extract_images(input_file: str, *args: str) -> List[str]:
         output_dir = mkdtemp()
         with TemporaryFilePath() as output_file_name:
             commands = [
@@ -202,3 +203,10 @@ class TestDumpImages:
         image_files = self.extract_images(filepath)
         assert len(image_files) == 23
         assert all(x.endswith(".bmp") for x in image_files)
+
+    def test_contrib_issue_1057_tiff_predictor(self) -> None:
+        """Test for extracting tiff image"""
+        filepath = absolute_sample_path("contrib/issue-1057-tiff-predictor.pdf")
+        image_files = self.extract_images(filepath)
+        assert len(image_files) == 1
+        assert image_files[0].endswith(".bmp")